amplify/backend/api/test2/src/python/requirements.txt
yfinance==0.2.31
aws_wsgi==0.2.7
boto3==1.28.62
Flask==1.1.4
Flask_Cors==4.0.0
Requests==2.31.0
MarkupSafe<2.0.0
python-dotenv==1.0.0
requests-html==0.10.0
python-math==0.0.1
